requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.3.0
matplotlib==3.8.4
seaborn==0.13.2
openpyxl==3.1.2
# WeasyPrint==62.1  # Comentado temporalmente por problemas en Windows
django-crispy-forms==2.1
crispy-bootstrap5==2024.2